Rename files, add class that is wrapper for working with dictionary
word_analyzer.py was changed to main.py because that's more expedient. Now word_analyzer.py provides the class-wrapper for working with dictionary.
WordAnalyzer has get_total_cost function which returns the total sum of the cost of each word in the sentence without any spaces. Zipf's law is used to calculate the word cost.