Add class for splitting the text without spaces
The class builds dictionary with pairs (word, word cost). The word is selected from the passed list of words into constructor of that class. Word cost is evaluated based on the order number in the list of words (remember, that list of words is sorted by frequency usage). Max_len was defined to detect a possibly word match.

The split function implements splitting the text. At first, build the cost array. It's built using best_match function that is defining the best match for i index, assuming cost has been build for i-1 indices.