Correct tests for arguments_parser function
Remove blocks where incorrect inputs in the command line were checked due to the fact that it's unnecessary, because if system error is raised, output is printed to stdout, i.e. to command line.

Also arguments of encoding and tree filename were be added and tests were be added for them.