Edit arguments of the ArgParser to make 4 modes of the program, edit the getting a mode method, add the getting verbose method
Need to make 4 modes of the program, so arguments of the ArgParser were be changed to they get 1 or no one parameter. The getting mode method were be edit, too, because of the mode is a dictionary now. The key of this dictionary is a mode and the value is a filename which a user specified after an option.
A verbose flag is not a mode now. It was changed and moved to the particular method of getting verbose.